+++ app/[lang]/(app)/(home)/(tablet)/restaurant/new-order/components/order-info/OrderInfo.tsx
@@ -873,7 +873,9 @@ export default function OrderInfo({ dic }: { dic: NewOrderDictionary }) {
          <NumericFormat
           {...other}
           value={value}
-          onValueChange={({ floatValue }) => onChange(floatValue || '')}
+          onValueChange={({ floatValue }) =>
+           onChange(floatValue || floatValue === 0 ? floatValue : '')
+          }
           id='discount-rate'
           customInput={InputGroupInput}
           allowNegative={false}